Moved api and revoke related methods to revoke.js reducer and saga
