+++ src/reducers/index.js
@@ -2,10 +2,10 @@ import { combineReducers } from "redux";
 
 import application from "./application";
 import admin from "./admin";
-import api from "./api";
+import revoke from "./revoke";
 
 export default combineReducers({
   application,
   admin,
-  api
+  revoke
 });
